refactor: mv controller types to a separate file
